+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased changes
 
+- Implement Serial for `num::rational::Ratio<u64>` and `Duration`.
+- Introduce types for protocol version 6: `ChainParameterVersion2`, `GASRewardsV1`, `TimeoutParameters` and `FinalizationCommitteeParameters`.
+
 ## 1.2.0 (2023-05-08)
 
 - Add helpers to extract policy from credentials.