+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased changes
 
 - Implement Serial for `num::rational::Ratio<u64>` and `Duration`.
-- Introduce types for protocol version 6: `ChainParameterVersion2`, `GASRewardsV1`, `TimeoutParameters` and `FinalizationCommitteeParameters`.
+- Introduce types for protocol version 6: `ChainParameterVersion2`, `GASRewardsCPV1`, `TimeoutParameters` and `FinalizationCommitteeParameters`.
 
 ## 1.2.0 (2023-05-08)