+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased changes
 
+- Extend types `UpdatePayload` and `UpdateType` with variants introduced in protocol version 6.
+- Implement `Serial` and `Deserial` for `num::rational::Ratio<u64>` and `Duration`.
+- Introduce types for protocol version 6: `Ratio`, `ChainParameterVersion2`, `GASRewardsCPV1`, `TimeoutParameters` and `FinalizationCommitteeParameters`.
+
 ## 1.2.0 (2023-05-08)
 
 - Add helpers to extract policy from credentials.